Frontend/requirements.txt
streamlit
pandas
request
json
os